[shared-ui] Move Component Selector to overlay
