+++ packages/shared-ui/src/elements/editor/ribbon.ts
@@ -37,6 +37,7 @@ import { NodeSelector } from "./node-selector";
 import { InspectableGraph, SubGraphs } from "@google-labs/breadboard";
 import { classMap } from "lit/directives/class-map.js";
 import { MAIN_BOARD_ID } from "../../constants/constants";
+import { guard } from "lit/directives/guard.js";
 
 const COLLAPSED_MENU_BUFFER = 60;
 
@@ -168,11 +169,14 @@ export class RibbonMenu extends LitElement {
       margin-right: 0;
     }
 
-    bb-node-selector {
+    bb-component-selector-overlay {
       display: none;
       position: absolute;
-      top: calc(100% + 10px);
-      height: calc(100svh - 122px);
+    }
+
+    :host([showcomponentselector="true"]) bb-component-selector-overlay {
+      display: block;
+      pointer-events: auto;
     }
 
     #save {
@@ -207,11 +211,6 @@ export class RibbonMenu extends LitElement {
       margin: 0 var(--bb-grid-size-4) 0 0;
       height: var(--bb-grid-size-5);
       border-left: 1px solid var(--bb-neutral-300);
-    }
-
-    :host([showcomponentselector="true"]) bb-node-selector {
-      display: grid;
-      pointer-events: auto;
     }
 
     #component-toggle {
@@ -727,25 +726,29 @@ export class RibbonMenu extends LitElement {
   }
 
   render() {
+    const componentSelector: HTMLTemplateResult = html`${guard(
+      [],
+      () =>
+        html`<bb-component-selector-overlay
+          .graph=${this.graph}
+          .showExperimentalComponents=${this.showExperimentalComponents}
+          @bbkitnodechosen=${(evt: KitNodeChosenEvent) => {
+            const id = createRandomID(evt.nodeType);
+            this.dispatchEvent(new NodeCreateEvent(id, evt.nodeType));
+          }}
+          @bboverlaydismissed=${() => {
+            this.showComponentSelector = false;
+          }}
+        ></bb-component-selector-overlay>`
+    )}`;
+
     const components = html`
-      <bb-node-selector
-        ${ref(this.#nodeSelectorRef)}
-        .graph=${this.graph}
-        .showExperimentalComponents=${this.showExperimentalComponents}
-        @bbkitnodechosen=${(evt: KitNodeChosenEvent) => {
-          const id = createRandomID(evt.nodeType);
-          this.dispatchEvent(new NodeCreateEvent(id, evt.nodeType));
-        }}
-        @bbdismissed=${() => {
-          this.showComponentSelector = false;
-        }}
-      ></bb-node-selector>
       <div id="component-toggle-container">
         <button
           id="component-toggle"
           class=${classMap({ active: this.showComponentSelector })}
           @click=${() => {
-            this.showComponentSelector = !this.showComponentSelector;
+            this.showComponentSelector = true;
           }}
         >
           Components
@@ -1274,6 +1277,7 @@ export class RibbonMenu extends LitElement {
     }
 
     const left = [
+      componentSelector,
       components,
       boardManagement,
       editControls,
