+++ packages/shared-ui/src/elements/elements.ts
@@ -16,6 +16,7 @@ export { ConnectionBroker } from "./connection/connection-broker.js";
 export { ConnectionInput } from "./connection/connection-input.js";
 export { ConnectionSettings } from "./connection/connection-settings.js";
 export { ConnectionSignin } from "./connection/connection-signin.js";
+export { ComponentSelectorOverlay } from "./overlay/component-selector.js";
 export { CommentOverlay } from "./overlay/comment.js";
 export { DragDockOverlay } from "./overlay/drag-dock-overlay.js";
 export { DrawableInput } from "./input/drawable/drawable.js";
@@ -38,7 +39,6 @@ export { NodeConfigurationOverlay } from "./overlay/node-configurator.js";
 export { NodeMetaDetails } from "./node-info/node-meta-details.js";
 export { NodeConfigurationInfo } from "./node-info/node-configuration.js";
 export { NodeRunner } from "./node-runner/node-runner.js";
-export { NodeSelector } from "./editor/node-selector.js";
 export { OverflowMenu } from "./overflow-menu/overflow-menu.js";
 export { Overlay } from "./overlay/overlay.js";
 export { PortTooltip } from "./editor/port-tooltip.js";
